Give every codec scenario a default pipeline
Extend the decoder-MD5 dispatch shape (generic default + optional
codec_<family>.py override via getattr) to all six scenarios:

- encoder PSNR gains a GenericEncoderProject reference pipeline
  (decodebin -> videoconvert -> encoder -> mux) used when a platform
  ships no create_encoder_psnr_project
- the transform pipelines move back into the scenario scripts as
  Generic*Project defaults - v4l2convert is a generic V4L2 element, so
  genio needs no transform hooks anymore (pipelines byte-identical)
- AV-sync gains the platform from the resource (-p, additive record
  field) and an optional build_audio_video_sync_command hook

Command-only scenarios use build_<scenario>_command hooks; scenarios
that validate artifacts use create_<scenario>_project hooks. Every
hook is optional now; job ids are unchanged.
