+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/bin/codec_genio.py
@@ -29,9 +29,7 @@ from gst_utils import (
     GST_LAUNCH_BIN,
     GStreamerMuxerType,
     GStreamerEncodePlugins,
-    GStreamerTransformActions,
     generate_artifact_name,
-    get_big_bug_bunny_golden_sample,
     get_test_file_path_by_params,
 )
 
@@ -46,31 +44,6 @@ def create_encoder_psnr_project(args: argparse.Namespace):
         height=args.height,
         framerate=args.framerate,
         mux=args.mux,
-    )
-
-
-def create_transform_rotate_and_flip_project(args: argparse.Namespace):
-    """Create the rotate/flip transform project for Genio."""
-    return GenioTransformRotateAndFlipProject(
-        platform=args.platform,
-        codec=args.encoder_plugin,
-        action=args.action,
-        width=args.width,
-        height=args.height,
-        framerate=args.framerate,
-    )
-
-
-def create_transform_resize_project(args: argparse.Namespace):
-    """Create the resize transform project for Genio."""
-    return GenioTransformResizeProject(
-        platform=args.platform,
-        codec=args.encoder_plugin,
-        width_from=args.width_from,
-        height_from=args.height_from,
-        width_to=args.width_to,
-        height_to=args.height_to,
-        framerate=args.framerate,
     )
 
 
@@ -206,171 +179,3 @@ class GenioProject(BaseCodecProject):
             self.artifact_file,
         )
         return final_pipeline
-
-
-class GenioTransformRotateAndFlipProject(BaseCodecProject):
-    """Genio rotate/flip transform pipeline handler and builder."""
-
-    def __init__(
-        self,
-        platform: str,
-        codec: str,
-        action: GStreamerTransformActions,
-        width: int,
-        height: int,
-        framerate: int,
-    ) -> None:
-        super().__init__(
-            platform=platform,
-            codec=codec,
-            width=width,
-            height=height,
-            framerate=framerate,
-        )
-        self._action = action
-        self._codec_parser_map = {
-            GStreamerEncodePlugins.V4L2H264ENC.value: "h264parse"
-        }
-        self._actions_map = {
-            GStreamerTransformActions.ROTATE_90: "rotate=90",
-            GStreamerTransformActions.ROTATE_180: "rotate=180",
-            GStreamerTransformActions.ROTATE_270: "rotate=270",
-            GStreamerTransformActions.HORIZONTAL_FLIP: "horizontal_flip=1",
-            GStreamerTransformActions.VERTICAL_FLIP: "vertical_flip=1",
-        }
-        # This sample video file will be consumed by any gstreamer piple as
-        # input video.
-        self._golden_sample = get_big_bug_bunny_golden_sample(
-            width=self._width, height=self._height, framerate=self._framerate
-        )
-        self._pipeline_builders = {
-            GStreamerEncodePlugins.V4L2H264ENC.value: (
-                self._transform_pipeline_builder
-            ),
-        }
-
-    @property
-    def psnr_reference_file(self) -> str:
-        """
-        A golden reference which has been transformed in advance. It's used to
-        be the compared reference file for PSNR.
-        """
-        golden_reference = "big_bug_bunny_{}x{}_{}fps_{}.mp4".format(
-            self._width, self._height, self._framerate, self._action
-        )
-
-        full_path = os.path.join(
-            VIDEO_CODEC_TESTING_DATA, SAMPLE_2_FOLDER, golden_reference
-        )
-        if not os.path.exists(full_path):
-            raise SystemExit(
-                "Error: Golden PSNR reference '{}' doesn't exist".format(
-                    full_path
-                )
-            )
-
-        return full_path
-
-    def _transform_pipeline_builder(self) -> str:
-        """
-        Build the gstreamer pipeline performing the rotate/flip action.
-        """
-        pipeline = (
-            "{} filesrc location={} ! decodebin ! v4l2convert "
-            "extra-controls='cid,{}'"
-        ).format(
-            GST_LAUNCH_BIN,
-            self._golden_sample,
-            self._actions_map.get(self._action),
-        )
-
-        if self._action in [
-            GStreamerTransformActions.ROTATE_90,
-            GStreamerTransformActions.ROTATE_270,
-        ]:
-            pipeline = (
-                "{} ! video/x-raw,width={},height={},"
-                "pixel-aspect-ratio='(fraction)1/1'"
-            ).format(pipeline, self._height, self._width)
-
-        pipeline = ("{} ! {} ! {} ! mp4mux ! filesink location={}").format(
-            pipeline,
-            self._codec,
-            self._codec_parser_map.get(self._codec),
-            self.artifact_file,
-        )
-        return pipeline
-
-
-class GenioTransformResizeProject(BaseCodecProject):
-    """Genio resize transform pipeline handler and builder."""
-
-    def __init__(
-        self,
-        platform: str,
-        codec: str,
-        width_from: int,
-        height_from: int,
-        width_to: int,
-        height_to: int,
-        framerate: int,
-    ) -> None:
-        super().__init__(
-            platform=platform,
-            codec=codec,
-            width=width_from,
-            height=height_from,
-            framerate=framerate,
-        )
-        self._width_to = width_to
-        self._height_to = height_to
-        self._codec_parser_map = {
-            GStreamerEncodePlugins.V4L2H264ENC.value: "h264parse"
-        }
-        # This sample video file will be consumed by any gstreamer piple as
-        # input video.
-        self._golden_sample = get_test_file_path_by_params(
-            self._width, self._height, self._framerate, self._codec
-        )
-        self._pipeline_builders = {
-            GStreamerEncodePlugins.V4L2H264ENC.value: (
-                self._resize_pipeline_builder
-            ),
-        }
-
-    @property
-    def psnr_reference_file(self) -> str:
-        """
-        A golden reference which has been transformed in advance. It's used to
-        be the compared reference file for PSNR.
-        """
-        golden_reference = get_test_file_path_by_params(
-            self._width_to, self._height_to, self._framerate, self._codec
-        )
-        if not os.path.exists(golden_reference):
-            raise SystemExit(
-                "Error: Golden PSNR reference '{}' doesn't exist".format(
-                    golden_reference
-                )
-            )
-
-        return golden_reference
-
-    def _resize_pipeline_builder(self) -> str:
-        """
-        Build the gstreamer pipeline scaling the stream while encoding.
-        """
-        pipeline = (
-            "{} filesrc location={} ! decodebin ! v4l2convert ! "
-            "video/x-raw,width={},height={} ! {} ! {} ! mp4mux ! filesink"
-            " location={}"
-        ).format(
-            GST_LAUNCH_BIN,
-            self._golden_sample,
-            self._width_to,
-            self._height_to,
-            self._codec,
-            self._codec_parser_map.get(self._codec),
-            self.artifact_file,
-        )
-        return pipeline
