+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/bin/gst_resources_generator.py
@@ -252,6 +252,7 @@ class GstResources:
                             sample_file["file_name"],
                             video_sink,
                         ),
+                        "platform": self._conf_name,
                         "video_sink": video_sink,
                         "decoder_plugin": item["decoder_plugin"],
                         "golden_sample_file_name": sample_file["file_name"],
