+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/units/video-codec/jobs.pxu
@@ -80,7 +80,7 @@ _verification:
     1. Do you hear the beep sound and see the video be displayed?
     2. Is the timing of audio and video synchronized?
 command:
-    gst_v4l2_audio_video_synchronization.py -dp {{decoder_plugin}} -gp {{golden_sample_file}} -vs {{video_sink}} -cp "{{capssetter_pipeline}}"
+    gst_v4l2_audio_video_synchronization.py -dp {{decoder_plugin}} -gp {{golden_sample_file}} -vs {{video_sink}} -cp "{{capssetter_pipeline}}" -p "{{ platform }}"
 
 unit: template
 template-engine: jinja2
